requirements.txt
Django~=4.0.1
loguru~=0.5.3
requests~=2.27.1
psycopg2~=2.9.3
redis~=4.1.1
celery~=5.2.3
djangorestframework~=3.12.4
https://github.com/mher/flower/zipball/master#egg=flower

